migrate TripCard to styled-components
